scraper/requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0

# --- Scraping ---
playwright==1.48.0
beautifulsoup4==4.12.3
requests==2.32.3

# --- PDF Extraction ---
pymupdf==1.24.10        # Best for PDF text extraction
pdfminer.six==20240706  # Backup extractor

# --- OCR (for scanned PDFs) ---
pytesseract==0.3.10
Pillow==10.4.0

# --- LLM / AI ---
google-generativeai==0.3.0

# --- PPT Generation ---
python-pptx==0.6.23

# --- Scheduling (Monthly Jobs) ---
schedule==1.2.2         # For cron-like Python jobs

# --- Utilities ---
python-dotenv==1.0.1    # Load .env config
loguru==0.7.2           # Beautiful logging
tqdm==4.66.4            # Progress bars
pandas==2.2.3           # Optional: table extraction support

# --- Optional (if using Celery/RQ for background tasks) ---
# celery==5.4.0
# redis==5.0.8
